+++ internal/grpcwrapper/rawydb/issues.go
@@ -17,7 +17,7 @@ func (issuesPointer *Issues) FromProto(p []*Ydb_Issue.IssueMessage) error {
 	issues := *issuesPointer
 	for i := range issues {
 		if err := issues[i].FromProto(p[i]); err != nil {
-			return nil
+			return err
 		}
 	}
 	return nil