+++ log/retry.go
@@ -13,7 +13,7 @@ func Retry(l Logger, d trace.Detailer, opts ...Option) (t trace.Retry) {
 	return internalRetry(wrapLogger(l, opts...), d)
 }
 
-func internalRetry(l *wrapper, d trace.Detailer) (t trace.Retry) {
+func internalRetry(l Logger, d trace.Detailer) (t trace.Retry) {
 	t.OnRetry = func(
 		info trace.RetryLoopStartInfo,
 	) func(