+++ tests/integration/topic_read_writer_test.go
@@ -163,8 +163,6 @@ func TestMessageMetadata(t *testing.T) {
 }
 
 func TestManyConcurentReadersWriters(t *testing.T) {
-	xtest.AllowByFlag(t, "ISSUE-389")
-
 	const partitionCount = 3
 	const writersCount = 5
 	const readersCount = 10
@@ -351,8 +349,6 @@ func TestCommitUnexpectedRange(t *testing.T) {
 }
 
 func TestUpdateToken(t *testing.T) {
-	xtest.AllowByFlag(t, "LOGBROKER-7960")
-
 	ctx := context.Background()
 	db := connect(t)
 	dbLogging := connectWithGrpcLogging(t)